修复一些启动项问题，完善分区自动扩展

+++ overlay/etc/init.d/rockchip.sh
@@ -39,7 +39,11 @@ install_mali() {
             ;;
     esac
 
-    apt install -f /packages/libmali/libmali-*$MALI*-x11*.deb
+    if [ ! -e /packages/libmali/libmali-*$MALI*-x11*.deb ]; then
+        echo "No libmali-*$MALI*-x11*.deb ."
+    else
+        apt install -f /packages/libmali/libmali-*$MALI*-x11*.deb
+    fi    
 }
 
 
@@ -88,23 +92,25 @@ then
     mount -o remount,sync /
 
     install_mali ${CHIPNAME}
-    setcap CAP_SYS_ADMIN+ep /usr/bin/gst-launch-1.0
-
-    # Cannot open pixbuf loader module file
-    if [ -e "/usr/lib/arm-linux-gnueabihf" ] ;
-    then
-       /usr/lib/arm-linux-gnueabihf/gdk-pixbuf-2.0/gdk-pixbuf-query-loaders > /usr/lib/arm-linux-gnueabihf/gdk-pixbuf-2.0/2.10.0/loaders.cache
-       update-mime-database /usr/share/mime/
-    elif [ -e "/usr/lib/aarch64-linux-gnu" ];
-    then
-       /usr/lib/aarch64-linux-gnu/gdk-pixbuf-2.0/gdk-pixbuf-query-loaders > /usr/lib/aarch64-linux-gnu/gdk-pixbuf-2.0/2.10.0/loaders.cache
-    fi
 
     rm -rf /packages
 
-    # The base target does not come with lightdm
-    systemctl restart lightdm.service || true
+    if [ -e /usr/bin/gst-launch-1.0 ]; then
+        setcap CAP_SYS_ADMIN+ep /usr/bin/gst-launch-1.0 
 
+        # Cannot open pixbuf loader module file
+        if [ -e "/usr/lib/arm-linux-gnueabihf" ] ;
+        then
+        /usr/lib/arm-linux-gnueabihf/gdk-pixbuf-2.0/gdk-pixbuf-query-loaders > /usr/lib/arm-linux-gnueabihf/gdk-pixbuf-2.0/2.10.0/loaders.cache
+        update-mime-database /usr/share/mime/
+        elif [ -e "/usr/lib/aarch64-linux-gnu" ];
+        then
+        /usr/lib/aarch64-linux-gnu/gdk-pixbuf-2.0/gdk-pixbuf-query-loaders > /usr/lib/aarch64-linux-gnu/gdk-pixbuf-2.0/2.10.0/loaders.cache
+        fi
+
+        # The base target does not come with lightdm
+        systemctl restart lightdm.service || true
+    fi
     touch /usr/local/first_boot_flag
 fi
